+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
 
+    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
+
     <application
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
@@ -12,6 +15,23 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SamdollarFront"
         tools:targetApi="31">
+
+        <!--
+             TODO: Before you run your application, you need a Google Maps API key.
+
+             To get one, follow the directions here:
+
+                https://developers.google.com/maps/documentation/android-sdk/get-api-key
+
+             Once you have your API key (it starts with "AIza"), define a new property in your
+             project's local.properties file (e.g. MAPS_API_KEY=Aiza...), and replace the
+             "YOUR_API_KEY" string in this file with "${MAPS_API_KEY}".
+        -->
+        <meta-data
+            android:name="com.google.android.geo.API_KEY"
+            android:value="AIzaSyCdDq1yU9MV05DaDEP2WhIwEwlnl0d9T2g" />
+
+
         <activity
             android:name=".MenuActivity"
             android:exported="false" />
@@ -28,8 +48,9 @@
             android:name=".MineActivity"
             android:exported="false" />
         <activity
-            android:name=".MainPageActivity"
-            android:exported="true" >
+            android:name=".MainActivity"
+            android:exported="true"
+            android:label="@string/title_activity_main" >
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
