+++ app/src/main/java/com/example/samdollarfront/MainActivity.kt
@@ -1,0 +1,63 @@
+package com.example.samdollarfront
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.widget.ImageButton
+
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.OnMapReadyCallback
+import com.google.android.gms.maps.SupportMapFragment
+import com.google.android.gms.maps.model.LatLng
+import com.google.android.gms.maps.model.MarkerOptions
+import com.example.samdollarfront.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity(), OnMapReadyCallback {
+
+    private lateinit var mMap: GoogleMap
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        val buttontoOwner = findViewById<ImageButton>(R.id.btn_owner)
+        buttontoOwner.setOnClickListener {
+            val intent1 = Intent(this, OwnerActivity::class.java)
+            startActivity(intent1)
+        }
+
+        val buttontoZzim = findViewById<ImageButton>(R.id.btn_zzim)
+        buttontoZzim.setOnClickListener {
+            val intent2 = Intent(this, ZzimActivity::class.java)
+            startActivity(intent2)
+        }
+
+        // Obtain the SupportMapFragment and get notified when the map is ready to be used.
+        val mapFragment = supportFragmentManager
+            .findFragmentById(R.id.map) as SupportMapFragment
+        mapFragment.getMapAsync(this)
+    }
+
+    /**
+     * Manipulates the map once available.
+     * This callback is triggered when the map is ready to be used.
+     * This is where we can add markers or lines, add listeners or move the camera. In this case,
+     * we just add a marker near Sydney, Australia.
+     * If Google Play services is not installed on the device, the user will be prompted to install
+     * it inside the SupportMapFragment. This method will only be triggered once the user has
+     * installed Google Play services and returned to the app.
+     */
+    override fun onMapReady(googleMap: GoogleMap) {
+        mMap = googleMap
+
+        // Add a marker in Sydney and move the camera
+        val marker = LatLng(35.241615, 128.695587)
+        mMap.addMarker(MarkerOptions().position(marker).title("Marker in Seoul"))
+        mMap.moveCamera(CameraUpdateFactory.newLatLng(marker))
+    }
+
+}
