+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">SamdollarFront</string>
+    <string name="title_activity_main">MainActivity</string>
 </resources>
